requirements.txt
Flask==2.3.3
google-ads==22.0.0
PyYAML==6.0.1